#38. Translate numbers with fractional part.

+++ spec/numbers_and_words/numbers_and_words_spec.rb
@@ -2,7 +2,7 @@
 require 'spec_helper'
 require 'shared_examples/correct_fixture_examples'
 
-[Integer, Array].each do |type|
+[Numeric, Array].each do |type|
   describe type do
     NumbersAndWords::I18n::Initialization.languages.each do |locale|
       context locale do
